+++ src/main.py
@@ -44,9 +44,10 @@ def main():
 
         topMargin = 0 if (dpi == 96) else int(10 * (dpi / 96))
         window.resize(
-            window.width() *
-            (dpi / 96), window.height() *
-            (dpi / 96)
+            int(window.width() *
+            (dpi / 96)),
+            int(window.height() *
+            (dpi / 96))
         )
         # window.verticalLayout_2.setContentsMargins(0, topMargin, 0, 0)
